Move LoginForm into login-page.tsx.
